refactor: extract custom radio button
